requirements.txt
allure_python_commons==2.5.5
logzero==1.5.0
pytest==3.7.0
parameterized==0.6.1
Pillow==5.4.1
selenium==3.141.0
pytest-allure-adaptor
pytest-ordering
pytest-rerunfailures
parameterize
pillow